+++ attributes/default.rb
@@ -1,2 +1,3 @@
-default['selinux']['state'] = 'enforcing'
 default['selinux']['booleans'] = {}
+default['selinux']['needs_reboot'] = nil
+default['selinux']['state'] = 'enforcing'